requirements.txt
absl-py==0.5.0
astor==0.7.1
certifi==2018.8.24
chardet==3.0.4
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
idna==2.7
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
Markdown==3.0
numpy==1.14.5
panda==0.3.1
protobuf==3.6.1
PyYAML==3.13
requests==2.19.1
scipy==1.1.0
six==1.11.0
tensorboard==1.10.0
tensorflow==1.10.1
termcolor==1.1.0
urllib3==1.23
Werkzeug==0.14.1
